Fix incorrect signer issue
