+++ FRW.xcodeproj/project.xcworkspace/xcshareddata/swiftpm/Package.resolved
@@ -141,7 +141,7 @@
         "repositoryURL": "https://github.com/Outblock/flow-swift",
         "state": {
           "branch": null,
-          "revision": "e76a25a287f89523a085ea7662d03b8bdf60af16",
+          "revision": "e7a1764953777a76858bdb373e2ae4f207bf5191",
           "version": null
         }
       },
