+++ FRW/Modules/Wallet/Send/WalletSendButtonView.swift
@@ -140,17 +140,6 @@ class MyAnimations: NSObject, ObservableObject {
     private var displaylink: CADisplayLink! // << here !!
 }
 
-// MARK: - ScaleButtonStyle
-
-struct ScaleButtonStyle: SwiftUI.ButtonStyle {
-    func makeBody(configuration: Self.Configuration) -> some View {
-        configuration.label
-            .scaleEffect(configuration.isPressed ? 0.95 : 1)
-            .animation(.linear(duration: 0.2), value: configuration.isPressed)
-            .brightness(configuration.isPressed ? -0.05 : 0)
-    }
-}
-
 // MARK: - WalletSendButtonView_Previews
 
 struct WalletSendButtonView_Previews: PreviewProvider {
