+++ FRW/Services/Manager/Wallet/WalletManager+Crypto.swift
@@ -70,46 +70,3 @@ extension WalletManager {
         return try cipher.decrypt(combinedData: data)
     }
 }
-
-// MARK: FlowSigner
-
-extension WalletManager: FlowSigner {
-    var keyIndex: Int {
-        currentMainAccount?.keyIndex ?? 0
-    }
-    
-    public var address: Flow.Address {
-        currentMainAccount?.address ?? .init(hex: "")
-    }
-
-    public func sign(transaction _: Flow.Transaction, signableData: Data) async throws -> Data {
-        return try await sign(signableData: signableData)
-    }
-
-    public func sign(signableData: Data) async throws -> Data {
-        let result = await SecurityManager.shared.SecurityVerify()
-        if result == false {
-            HUD.error(title: "verify_failed".localized)
-            throw WalletError.securityVerifyFailed
-        }
-        
-        guard let provider = keyProvider else {
-            throw WalletError.emptyKeyProvider
-        }
-        
-        guard let key = currentMainAccount?.fullWeightKey else {
-            throw WalletError.emptyAccountKey
-        }
-        
-        let signature = try provider.sign(
-            data: signableData,
-            signAlgo: key.signAlgo,
-            hashAlgo: key.hashAlgo
-        )
-        return signature
-    }
-
-    private func userSecretSign() -> Bool {
-        UserManager.shared.userType != .phrase
-    }
-}
